+++ src/MobileToken.ts
@@ -14,12 +14,10 @@
 // and limitations under the License.
 //
 
-import { type MobileTokenOperation } from './operations/MobileTokenOperation';
-import { type MobileTokenUserOperation } from './operations/MobileTokenUserOperation';
-import { PowerAuth, PowerAuthAuthentication } from 'react-native-powerauth-mobile-sdk';
-import { Platform } from 'react-native';
-
-export type RequestProcessor = (name: RequestInit) => RequestInit;
+import { Operations } from './operations/Operations';
+import { PowerAuth } from 'react-native-powerauth-mobile-sdk';
+import { Push } from './push/Push';
+import { Inbox } from './inbox/Inbox';
 
 /**
  * MobileToken class exposes API that enables to fetch, authorize or reject basic
@@ -27,323 +25,41 @@ export type RequestProcessor = (name: RequestInit) => RequestInit;
  */
 export class MobileToken {
 
-  /**
-   * Accept language for the outgoing requests headers.
-   * Default value is "en".
-   * 
-   * Standard RFC "Accept-Language" https://tools.ietf.org/html/rfc7231#section-5.3.5
-   * Response texts are based on this setting. For example when "de" is set, server
-   * will return operation texts in german (if available).
-   */
-  acceptLanguage = "en";
+    operations: Operations;
+    push: Push;
+    inbox: Inbox;
 
-  private pa: PowerAuth;
-  private baseURL: string;
+    /**
+     * 
+     * @param powerAuth PowerAuth instance. Needs to be activated when calling any method of this class - othewise error will be thrown.
+     * @param baseURL BaseURL of the server. If not provided, same URL as for PowerAuth it used.
+     */
+    constructor(powerAuth: PowerAuth, baseURL?: string) {
 
-  /**
-   * 
-   * @param powerAuth PowerAuth instance. Needs to be activated when calling any method of this class - othewise error will be thrown.
-   * @param baseURL BaseURL of the server. If not provided, same URL as for PowerAuth it used.
-   */
-  constructor(powerAuth: PowerAuth, baseURL?: string) {
-    this.pa = powerAuth;
-    this.baseURL = baseURL ?? powerAuth.configuration?.baseEndpointUrl ?? ""
-    if (!this.baseURL.endsWith("/")) {
-      this.baseURL = this.baseURL + "/";
-    }
-  }
+        let url = baseURL ?? powerAuth.configuration?.baseEndpointUrl ?? ""
+        if (!url.endsWith("/")) {
+            url += "/";
+        }
 
-  /**
-   * Retrieves user operations from the server.
-   * 
-   * @param requestProcessor You may modify the request via this processor. It's highly recommended to only modify HTTP headers.
-   * @returns Server response (with list of operations).
-   */
-  async operationList(requestProcessor?: RequestProcessor): Promise<MobileTokenResponse<MobileTokenUserOperation[]>> {
-    return await this.postSignedWithToken<MobileTokenUserOperation[]>(
-      {},
-      PowerAuthAuthentication.possession(),
-      "api/auth/token/app/operation/list",
-      "possession_universal",
-      true,
-      requestProcessor
-    );
-  }
-
-  /**
-   * Retrieves operation detail based on operation ID
-   * 
-   * @param operationId ID of the operation
-   * @param requestProcessor You may modify the request via this processor. It's highly recommended to only modify HTTP headers.
-   * @returns Server response (with operation detail)
-   */
-  async operationDetail(operationId: string, requestProcessor?: RequestProcessor): Promise<MobileTokenResponse<MobileTokenUserOperation>> {
-    return await this.postSignedWithToken<MobileTokenUserOperation>(
-      { requestObject: { id: operationId } },
-      PowerAuthAuthentication.possession(),
-      "api/auth/token/app/operation/detail",
-      "possession_universal",
-      true,
-      requestProcessor
-    );
-  }
-
-  /**
-   * Authorize operation with given PowerAuth authentication object.
-   * 
-   * @param operation Operation to authorize
-   * @param authentication Authentication object
-   * @param requestProcessor You may modify the request via this processor. It's highly recommended to only modify HTTP headers.
-   * @returns Server response
-   */
-  async authorize(operation: MobileTokenOperation, authentication: PowerAuthAuthentication, requestProcessor?: RequestProcessor): Promise<MobileTokenResponse<void>> {
-    return await this.postSigned<void>(
-      { requestObject: { id: operation.id, data: operation.data } },
-      authentication,
-      "/api/auth/token/app/operation/authorize",
-      "/operation/authorize",
-      false,
-      requestProcessor
-    );
-  }
-
-  /**
-   * Reject operation with a reason.
-   * 
-   * @param operationId ID of the operation.
-   * @param reason Reason for the rejection.
-   * @param requestProcessor You may modify the request via this processor. It's highly recommended to only modify HTTP headers.
-   * @returns Server response
-   */
-  async reject(operationId: string, reason: "INCORRECT_DATA" | "UNEXPECTED_OPERATION" | "UNKNOWN" | string, requestProcessor?: RequestProcessor): Promise<MobileTokenResponse<void>> {
-    return await this.postSigned<void>(
-      { requestObject: { id: operationId, reason: reason } },
-      PowerAuthAuthentication.possession(),
-      "/api/auth/token/app/operation/cancel",
-      "/operation/cancel",
-      false,
-      requestProcessor
-    );
-  }
-
-  /** 
-   * Registers the given powerauth activation for push notifications.
-   * 
-   * @param token Push token
-   * @param platform ios, android or huawei. When not specified, `Platform.OS` is used to determine if ios or android will be used. There is currently no automatic huawei platform detection.
-   * @param requestProcessor You may modify the request via this processor. It's highly recommended to only modify HTTP headers.
-   * @returns Server response
-   */
-  async registerForPush(token: string, platform?: "ios" | "android" | "huawei", requestProcessor?: RequestProcessor): Promise<MobileTokenResponse<void>> {
-
-    if (platform == undefined) {
-      // only
-      platform = Platform.OS == "ios" ? "ios" : "android";
+        this.operations = new Operations(powerAuth, url)
+        this.push = new Push(powerAuth, url);
+        this.inbox = new Inbox(powerAuth, url);
     }
 
-    return await this.postSignedWithToken<void>(
-      { requestObject: { token: token, platform: platform } },
-      PowerAuthAuthentication.possession(),
-      "/api/push/device/register/token",
-      "possession_universal",
-      false,
-      requestProcessor
-    );
-  }
-
-  private async postSigned<T>(
-    requestData: any,
-    auth: PowerAuthAuthentication,
-    endpoindPath: string,
-    uriId: string,
-    returnDataExpected: boolean,
-    requestProcessor?: RequestProcessor
-  ): Promise<MobileTokenResponse<T>> {
-
-    let body = JSON.stringify(requestData);
-    let paHeader = await this.pa.requestSignature(auth, "POST", uriId, body);
-    let headers = new Headers();
-    headers.set(paHeader.key, paHeader.value);
-    return await this.post(JSON.stringify(requestData), endpoindPath, returnDataExpected, headers, requestProcessor);
-  }
-
-  private async postSignedWithToken<T>(
-    requestData: any,
-    auth: PowerAuthAuthentication,
-    endpoindPath: string,
-    tokenName: string,
-    returnDataExpected: boolean,
-    requestProcessor?: RequestProcessor
-  ): Promise<MobileTokenResponse<T>> {
-
-    let token = await this.pa.tokenStore.requestAccessToken(tokenName, auth);
-    let paHeader = await this.pa.tokenStore.generateHeaderForToken(token.tokenName);
-
-    let headers = new Headers();
-    headers.set(paHeader.key, paHeader.value);
-
-    return await this.post(JSON.stringify(requestData), endpoindPath, returnDataExpected, headers, requestProcessor);
-  }
-
-  private async post<T>(
-    requestSerialized: string,
-    endpoindPath: string,
-    returnDataExpected: boolean,
-    headers: Headers,
-    requestProcessor?: RequestProcessor
-  ): Promise<MobileTokenResponse<T>> {
-    let method = "POST";
-    let url = this.baseURL + endpoindPath;
-
-    let jsonType = "application/json";
-    headers.set("Accept", jsonType);
-    headers.set("Content-Type", jsonType);
-    headers.set("Accept-Language", this.acceptLanguage);
-    headers.set("User-Agent", "react-native-mtoken-sdk"); // TODO: improve!
-
-    let request: RequestInit = {
-      method: method,
-      headers: headers,
-      body: requestSerialized
+    /**
+     * Sets accept language for the outgoing requests headers for `operations`, `push` and `inbox` objects.
+     * 
+     * The value can be further modified in the each object individualy.
+     * 
+     * Default value is "en".
+     * 
+     * 
+     * Standard RFC "Accept-Language" https://tools.ietf.org/html/rfc7231#section-5.3.5
+     * Response texts are based on this setting. For example when "de" is set, server
+     * will return operation texts in german (if available).
+     */
+    setAcceptLanguage(lang: string) {
+        this.operations.acceptLanguage = lang;
+        this.push.acceptLanguage = lang;
     }
-
-    if (requestProcessor) {
-      request = requestProcessor(request);
-    }
-
-    let result = await fetch(url, request);
-    let responseBody = await result.text()
-    let response = JSON.parse(responseBody, (key: string, value: any) => {
-      if (key == "operationExpires" || key == "operationCreated") {
-        return new Date(value);
-      }
-      return value
-    }) as MobileTokenResponse<T>;
-
-    if (response.status == "ERROR") {
-      if (response.responseObject == undefined) {
-        throw new MobileTokenException("Error retrieved but no error data", { ...result })
-      }
-      response.responseError = response.responseObject as any
-      response.responseObject = undefined
-    }
-
-    if (response.status == "OK" && returnDataExpected && response.responseObject == undefined) {
-      throw new MobileTokenException("No data object retieved.", { ...result })
-    }
-
-    return response;
-  }
-}
-
-/** Possible logic errors during the API calls. */
-export class MobileTokenException {
-
-  description: string;
-  additionalData?: any;
-  
-  constructor(description: string, additionalData?: any) {
-    this.description = description;
-    this.additionalData = additionalData;
-  }
-}
-
-/** Response from the API. */
-export interface MobileTokenResponse<T> {
-  status: "OK" | "ERROR";
-  responseError?: MobileTokenResponseError;
-  responseObject?: T;
-} 
-
-/** Error object when error on the server happens. */
-export interface MobileTokenResponseError {
-  code: KnownRestApiError | string;
-  message: string;
-}
-
-/** Known PowerAuth server error codes. */
-export enum KnownRestApiError {
-    
-  // COMMON ERRORS
-  
-  /// When unexpected error happened. */
-  GenericError                     = "ERROR_GENERIC",
-  
-  /** General authentication failure (wrong password, wrong activation state, etc...) */
-  AuthenticationFailure            = "POWERAUTH_AUTH_FAIL",
-  /// Invalid request sent - missing request object in request */
-  InvalidRequest                   = "INVALID_REQUEST",
-  /// Activation is not valid (it is different from configured activation) */
-  InvalidActivation                = "INVALID_ACTIVATION",
-  /// Invalid application identifier is attempted for operation manipulation */
-  InvalidApplication               = "INVALID_APPLICATION",
-  /// Invalid operation identifier is attempted for operation manipulation */
-  InvalidOperation                 = "INVALID_OPERATION",
-  /// Error during activation */
-  ActivationError                  = "ERR_ACTIVATION",
-  /// Error in case that PowerAuth authentication fails */
-  AuthenticationError              = "ERR_AUTHENTICATION",
-  /// Error during secure vault unlocking */
-  SecureVaultError                 = "ERR_SECURE_VAULT",
-  /// Returned in case encryption or decryption fails */
-  EncryptionError                  = "ERR_ENCRYPTION",
-  
-  // PUSH ERRORS
-
-  /// Failed to register push notifications */
-  PushRegistrationFailed           = "PUSH_REGISTRATION_FAILED",
-
-  // OPERATIONS ERRORS
-  
-  /// Operation is already finished */
-  OperationAlreadyFinished         = "OPERATION_ALREADY_FINISHED",
-  /// Operation is already failed */
-  OperationAlreadyFailed           = "OPERATION_ALREADY_FAILED",
-  /// Operation is cancelled */
-  OperationAlreadyCancelled        = "OPERATION_ALREADY_CANCELED",
-  /// Operation is expired */
-  OperationExpired                 = "OPERATION_EXPIRED",
-  /// Operation authorization failed */
-  OperationFailed                  = "OPERATION_FAILED",
-
-  // Following errors in the current mtoken implementatation should not happen, so we're commenting them out to keep it simple
-
-  // // ACTIVATION SPAWN ERRORS
-  
-  // /// Unable to fetch activation code.
-  // ActivationCodeFailed             = "ACTIVATION_CODE_FAILED",
-  
-  // // IDENTITY ONBOARDING ERRORS
-  
-  // /// Onboarding process failed or failed to start
-  // OnboardingFailed                 = "ONBOARDING_FAILED",
-  // /// An onboarding process limit reached (e.g. too many reset attempts for identity verification or maximum error score exceeded).
-  // OnboardingLimitReached           = "ONBOARDING_PROCESS_LIMIT_REACHED",
-  // /// Too many attempts to start an onboarding process for a user.
-  // OnboardingTooManyProcesses       = "TOO_MANY_ONBOARDING_PROCESSES",
-  // /// Failed to resend onboarding OTP (probably requested too soon)
-  // OnboardingOtpFailed              = "ONBOARDING_OTP_FAILED",
-  // /// Document is invalid
-  // InvalidDocument                  = "INVALID_DOCUMENT",
-  // /// Document submit failed
-  // DocumentSubmitFailed             = "DOCUMENT_SUBMIT_FAILED",
-  // /// Identity verification failed
-  // IdentityVerificationFailed       = "IDENTITY_VERIFICATION_FAILED",
-  // /// Identity verification limit reached (e.g. exceeded number of upload attempts).
-  // IdentityVerificationLimitReached = "IDENTITY_VERIFICATION_LIMIT_REACHED",
-  // /// Verification of documents failed
-  // DocumentVerificationFailed       = "DOCUMENT_VERIFICATION_FAILED",
-  // /// Presence check failed
-  // PresenceCheckFailed              = "PRESENCE_CHECK_FAILED",
-  // /// Presence check is not enabled
-  // PresenceCheckNotEnabled          = "PRESENCE_CHECK_NOT_ENABLED",
-  // /// Maximum limit of presence check attempts was exceeded.
-  // PresenceCheckLimitEached         = "PRESENCE_CHECK_LIMIT_REACHED",
-  // /// Too many same requests
-  // TooManyRequests                  = "TOO_MANY_REQUESTS",
-  
-  // // OTHER
-  
-  // /// Communication with remote system failed
-  // RemoteCommunicationError         = "REMOTE_COMMUNICATION_ERROR"
 }
